+++ src/components/loans/LoanRow.jsx
@@ -1,4 +1,15 @@
 // src/components/loans/LoanRow.jsx
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+  AlertDialogTrigger,
+} from "@/components/ui/alert-dialog";
 import { TableCell, TableRow } from "../ui/table";
 import { Button } from "../ui/button";
 import { BookOpen, Trash2, CheckCircle } from "lucide-react";
@@ -22,8 +33,8 @@ export const LoanRow = ({ loan, onDelete, onReturn }) => {
       </TableCell>
       <TableCell className="px-4 py-3">
         <div className="flex items-center gap-3">
-          <div className="flex h-10 w-10 items-center justify-center rounded-lg bg-primary/10">
-            <BookOpen className="h-5 w-5 text-primary" />
+          <div className="flex h-10 w-10 items-center justify-center rounded-lg bg-[#E6F5FC]">
+            <BookOpen className="h-5 w-5 text-[#43B7EC]" />
           </div>
           <div className="flex flex-col">
             <span className="font-medium text-foreground">
@@ -64,14 +75,36 @@ export const LoanRow = ({ loan, onDelete, onReturn }) => {
               Kembalikan
             </Button>
           )}
-          <Button
-            variant="ghost"
-            size="sm"
-            onClick={() => onDelete(loan.id)}
-            className="h-8 w-8 p-0 text-muted-foreground hover:text-destructive hover:bg-destructive/10 cursor-pointer"
-          >
-            <Trash2 className="h-4 w-4" />
-          </Button>
+          <AlertDialog>
+            <AlertDialogTrigger asChild>
+              <Button
+                variant="ghost"
+                size="icon"
+                className="text-destructive hover:bg-destructive/10"
+              >
+                <Trash2 className="h-4 w-4" />
+              </Button>
+            </AlertDialogTrigger>
+
+            <AlertDialogContent>
+              <AlertDialogHeader>
+                <AlertDialogTitle>Hapus Siswa Pinjam Buku?</AlertDialogTitle>
+                <AlertDialogDescription>
+                 Apakah Anda yakin ingin menghapus data peminjaman nama <b>{loan.siswa?.name ?? "Unknown Student"}</b> ini? 
+                </AlertDialogDescription>
+              </AlertDialogHeader>
+
+              <AlertDialogFooter>
+                <AlertDialogCancel>Batal</AlertDialogCancel>
+                <AlertDialogAction
+                  className="bg-destructive text-white hover:bg-destructive/90"
+                  onClick={() => onDelete(loan.id)}
+                >
+                  Ya, Hapus
+                </AlertDialogAction>
+              </AlertDialogFooter>
+            </AlertDialogContent>
+          </AlertDialog>
         </div>
       </TableCell>
     </TableRow>
